Document isolating multi-host stacks
